% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bg_genes}
\alias{bg_genes}
\title{Set of background genes}
\format{A character vector containing 29125 gene names}
\usage{
bg_genes
}
\description{
A dataset containing a set of background mouse genes.
All protein coding genes in a certain release?
}
\examples{
head(bg_genes)
}
\keyword{datasets}
