% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrepresentation_test.r, R/utilities.R
\name{clean_text}
\alias{clean_text}
\title{Clean text}
\usage{
clean_text(text, chars = c("\\t", "\\r", "\\n", "\\\\", " ", "\\""),
  remove_empty = TRUE, remove_dup = TRUE, upper_case = TRUE)

clean_text(text, chars = c("\\t", "\\r", "\\n", "\\\\", " ", "\\""),
  remove_empty = TRUE, remove_dup = TRUE, upper_case = TRUE)
}
\arguments{
\item{text}{character vector}

\item{chars}{character vector of characters to remove from text}

\item{remove_empty}{remove empty elements from the vector.}

\item{remove_dup}{remove duplicated elements from the vector.}

\item{upper_case}{convert to upper case}

\item{text}{character vector}

\item{chars}{character vector of characters to remove from text}

\item{remove_empty}{remove empty elements from the vector.}

\item{remove_dup}{remove duplicated elements from the vector.}

\item{upper_case}{convert to upper case}
}
\value{
A character vector

A character vector
}
\description{
Remove erroneous text characters.

Remove erroneous text characters.
}
\examples{
clean_text("\\thello_world")
clean_text("\\thello_world", upper_case = FALSE)
clean_text("\\thello_world")
clean_text("\\thello_world", upper_case = FALSE)
}
