% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes_1}
\alias{genes_1}
\title{Set of mouse genes}
\format{A character vector containing 100 gene names}
\usage{
genes_1
}
\description{
A dataset containing a set of 100 mouse genes.
}
\examples{
x <- head(genes_1)
x
}
\keyword{datasets}
