% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_categories}
\alias{human_categories}
\title{Human gene ontology categories}
\format{A named list of vectors
\describe{
  \item{name}{name of functional category }
  \item{genes}{vector of genes in the category}
  ...
}}
\source{
\url{http://download.baderlab.org/EM_Genesets/current_release/Human/symbol/Human_GO_AllPathways_no_GO_iea_January_01_2020_symbol.gmt}
}
\usage{
human_categories
}
\description{
A list of functional categories from gene ontology and other sources, and the genes in each category.
This had been parsed from a gmt file.
The source file is updated monthly so this should be updated too.
}
\examples{
head(human_categories, n = 2)
}
\keyword{datasets}
