% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_categories}
\alias{mouse_categories}
\title{Mouse gene ontology categories}
\format{A named list of vectors
\describe{
  \item{name}{name of functional category }
  \item{genes}{vector of genes in the category}
  ...
}}
\source{
\url{http://download.baderlab.org/EM_Genesets/current_release/Mouse/symbol/Mouse_GO_AllPathways_no_GO_iea_October_01_2019_symbol.gmt}
}
\usage{
mouse_categories
}
\description{
A list of functional categories from gene ontology and other sources, and the genes in each category.
This had been parsed from a gmt file.
The source file is updated monthly so this should be updated too.
}
\examples{
head(mouse_categories, n=3)
}
\keyword{datasets}
