% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{remove_duplicates}
\usage{
remove_duplicates(x, column = 1)
}
\arguments{
\item{x}{vector or data frame to remove duplicates from}

\item{column}{If a data frame is passed in, the column that is used
for deduplicating (default = 1)}
}
\value{
A character vector or data frame.
}
\description{
remove duplicates from vector or data frame
}
\examples{
hello_string <- c("h", "e", "l", "l", "o")
hello_string
remove_duplicates(hello_string)

df <- data.frame(hello_string, 1:5)
df
remove_duplicates(df)
}
